// src/pages/inventory/transactions/sales/transaction-types.ts
export interface GridQtyFactors {
  id: string;
  slNo: number;
  width: number;
  height: number;
  nos: number;
  total: number;
}